refactor(checkbox): rename component for docs sake
