+++ src/components/Checkbox/index.js
@@ -1,2 +1,2 @@
-export { default as MCheckbox } from './src/CheckboxFormItem.vue';
+export { default as MCheckbox } from './src/Checkbox.vue';
 export { default as MCheckboxControl } from './src/CheckboxControl.vue';
